Update tanks.json
This is a breaking change.

The addition of the `greywater` tank type and the change from `water` to `fresh water` is motivated by the fact that some boats have rainwater collection systems. If I recall correctly, greywater is also available in the N2K spec.

The addition of the `name` field is explained inline.

The change from `tankType` to `type` reflects the naming convention used in other parts of the spec.

The other changes are just to standardize the formatting and provide reader clarity.