Add deltaToFull and fillIdentity
The way delta messages are mapped to full docs is needed all over the place and should be
far upstream - specification is pretty much as far as it can go. The codebase needs going over
to use deltaToFull functionality from the signalk-schema (which is a part of the specification)
module.

FillIdentity is needed to set the identity property from the key under vessels so that the
produced full doc is schema compliant.