There are too many levels in Navigation
With the changes to Active Route there was a new layer Waypoint created which is a layer too far. There were some values like Bearing to next Waypoint in the Active Route level and other values like XTE in the Waypoint level which did not make any sense. Need all of the next waypoint navigation data in the Active Route level and to get rid of the Waypoint level.