Fix patternProperties & definitions
Show patternProperties ('maps' if you will) as placeholder * in the path. A Signal K application
could use this as a regexp to match a path against the set of paths. This
commit also attempts to fix  by implementing lookups in the same schema file's
definitions segment and in the global definitions.json - works to a degree. Includes
also an updated keysiwithmetadata generated with this logic.