Move Properties out of `anyOf`
While I think I understand the intent here, this isn't really how
`anyOf` is meant to be used and it fails validation. However, I'm not
sure how to correctly specify that if one of the subset of optional
properties exist, none of the others in that subset should be present.