Add Transverse Water Speed and Leeway Angle
A new generation of electro magnetic transducers that can measure the boat's speed through the water both longitudinal and transverse will provide a view of the boat's "Leeway" that was previously difficult to accurately calculate. Signal K needs to support these new transducers which will output the VBW sentence or PGN 130578.