Expand on the Discovery Section
Clarify mDNS/Bonjour and explain that the process being used is DNS-SD
via mDNS. This should give anyone who needs to implement this on a
platform without a ready-to-hand mDNS client library to get started in
the right direction.

Describe the fields or attributes which MUST, SHOULD and MAY be present
in the TXT record of the service advertisements and update the examples
to match.

Add a link to the section of the documentation which explains the
endpoints API and remove the example endpoint JSON as it is already
described in the other section.