Remove Service Identifiers and Rename TXT Fields
Remove specification for _https, _signalk_https and _signalk_wss. Add
language requiring a redirect to the secure versions of HTTP and
WebSockets if the server provides them.

Rename sever and version fields to swname and swvers.

Add language to specify the use of MMSI for self identifier if it is
present.