fix: don't create undefined meta properties
Before this change the full tree had undefined meta properties for example
under notifications, where the path is dynamic and no metadata is
available in the schema for values' paths. Undefined properties
do not matter in creating JSON, as they are ignored, but matter in
validation, breaking schema validation in a way that is very hard
to detect.