chore: remove dependency on tv4-formats
TV4-Format provides some JSON validation helper for common string
formats. It was loaded only when validating the full model (not delta)
and all the tests run without it so I propose we remove it.

Reason for me looking into this is that this library causes problem with
webpack because of a heuristic they use to require one of their
dependency. This is a known issue that they have fixed in a later
version and I have confirmed that upgrading the package fixes the issue
for us too. However, if we do not use the library I think it's better to
completely remove it.

For more info: https://github.com/ikr/tv4-formats/commit/61450dc7499440980169a67bc51d554a67c5c30d