test: make sure timestamp validation is working
Add two tests to make sure that timestamp validation is working.

The first test 'ending with Z' will verify that a timestamp not ending
with Z is rejected. This is the signalk spec regexp '.*Z'.

The second test is more subtle and requires the date to be an actually
valid timetamp. We currently use tv4-formats for this.